Optimize SQL explain steps in other flink version
